+++ src/components/Main/Cart/CartDescription/CartDescription.style.ts
@@ -1,0 +1,7 @@
+import { css } from "@emotion/react";
+
+export const CartItemTypesCountStyle = css({
+  fontSize: "12px",
+  fontWeight: "500",
+  textAlign: "left",
+});
