+++ src/components/Main/Cart/CartTitle/CartTitle.style.ts
@@ -7,9 +7,3 @@ export const CartTitleStyle = css({
 
   marginBottom: "12px",
 });
-
-export const CartItemTypesCountStyle = css({
-  fontSize: "12px",
-  fontWeight: "500",
-  textAlign: "left",
-});
