+++ src/components/Main/Cart/CartTitle/CartTitle.tsx
@@ -1,20 +1,14 @@
-import { useRecoilValue } from "recoil";
-import { CartItemTypesCountStyle, CartTitleStyle } from "./CartTitle.style";
-import { cartState } from "@/store/atom/atoms";
+import { CartTitleStyle } from "./CartTitle.style";
 import { PropsWithChildren } from "react";
 
 interface CartTitle extends PropsWithChildren {
   description?: string;
 }
 
-const CartTitle = ({ description, children }: CartTitle) => {
-  const itemCount = useRecoilValue(cartState).length;
-
+const CartTitle = ({ children }: CartTitle) => {
   return (
     <div>
       <h1 css={CartTitleStyle}>{children}</h1>
-      {itemCount !== 0 && <div css={CartItemTypesCountStyle}>현재 {itemCount}종류의 상품이 담겨 있습니다.</div>}
-      {description && <div css={CartItemTypesCountStyle}>{description}</div>}
     </div>
   );
 };
